requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cloudinary==1.33.0
dj-database-url==2.0.0
Django==4.2.2
gunicorn==20.1.0
Pillow==9.5.0
psycopg2==2.9.6
python-decouple==3.8
python-magic==0.4.27
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
whitenoise==6.5.0